Squashed commit of the following:

commit 4b0fe2fef20c5b1011e272a2e5e8f36fa5414fe4
Author: Martin Migasiewicz <martin.migasiewicz@tomtom.com>
Date:   Wed Jan 30 22:53:38 2019 +0100

    Rename nested fields to object so that the template won't map them to nested types.
    Added as well a test summary including total count of failed/passed/skipped tests

+++ elastic/hooks/quickbuild/store_results.py
@@ -100,7 +100,7 @@ def format_quickbuild_results(build_test_data):
                 'test': test_details[QBResultsExporter.KEY_TEST_NAME],
                 'result': test_result.name,
                 'message': test_details[QBResultsExporter.KEY_ERROR_MESSAGE],
-                'duration': duration,
+                'duration': float(duration),
                 'reportset': test_details[QBResultsExporter.KEY_REPORT_SET]
             }
 
@@ -123,7 +123,7 @@ def format_quickbuild_results(build_test_data):
                 suite_result['package'] = package
 
             try:
-                suite_result['duration'] += int(duration)
+                suite_result['duration'] += float(duration)
             except ValueError:
                 pass
